+++ lumium-space/src/sections/workspace/NavBar.tsx
@@ -1,22 +1,21 @@
 import { MoonIcon, SunIcon } from "@chakra-ui/icons";
-import { Box, Image, Text, FlexProps, Flex, IconButton, Stack, HStack, Button, Menu, MenuButton, Avatar, VStack, MenuList, MenuItem, MenuDivider } from "@chakra-ui/react";
-import { useColorMode, useColorModeValue } from "@chakra-ui/system";
+import { useColorMode, useColorModeValue, Box, Image, Text, FlexProps, Flex, IconButton, Stack, HStack, Button, Menu, MenuButton, Avatar, VStack, MenuList, MenuItem, MenuDivider, Spacer } from "@chakra-ui/react";
+import { useApi } from "@hooks/api";
+import { SECURE_AUTH_SIGNOUT } from "@routes/api/v1";
+import { ACCOUNT, ROOT } from "@routes/space";
 import { UserDTO, WorkspaceDTO } from "@types";
+import Router from "next/router";
 import { FiMenu, FiBell, FiChevronDown } from "react-icons/fi";
 
-interface MobileProps extends FlexProps {
+interface NavBarProps extends FlexProps {
     onOpen: () => void;
     userInfo?: UserDTO;
     workspace?: WorkspaceDTO;
 }
 
-const MobileNav = ({ onOpen, userInfo, workspace, ...rest }: MobileProps) => {
+const NavBar = ({ onOpen, userInfo, workspace, ...rest }: NavBarProps) => {
     const { colorMode, toggleColorMode } = useColorMode();
-
-    const darkLogo = '/logo/svg/Black logo - no background.svg';
-    const lightLogo = '/logo/svg/White logo - no background.svg';
-    let logo = useColorModeValue(darkLogo, lightLogo);
-
+    const [api] = useApi();
     let role = ""
     if (userInfo?.id) {
         if (workspace?.ownerId == userInfo?.id) {
@@ -32,27 +31,24 @@ const MobileNav = ({ onOpen, userInfo, workspace, ...rest }: MobileProps) => {
 
     return (
         <Flex
-            ml={{ base: 0, md: 60 }}
-            px={{ base: 4, md: 4 }}
-            height="20"
-            alignItems="center"
+            alignItems={"Center"}
             borderBottomWidth="1px"
             borderBottomColor={useColorModeValue('gray.200', 'gray.700')}
-            justifyContent={{ base: 'space-between', md: 'flex-end' }}
-            {...rest}>
+            {...rest}
+            pl={{ base: "4%", md: "2%" }}
+            pr={{ base: "4%", md: "2%" }}
+            pt={"1vh"}
+            pb={"1vh"}
+        >
             <IconButton
-                display={{ base: 'flex', md: 'none' }}
+                display={'flex'}
                 onClick={onOpen}
                 variant="outline"
                 aria-label="open menu"
                 icon={<FiMenu />}
             />
-
-            <Stack align={'center'} display={{ base: 'flex', md: 'none' }}>
-                <Image src={logo} minWidth={"70%"} maxWidth={"80%"} maxH={20} alt="lumium logo" />
-            </Stack>
-
-            <HStack spacing={{ base: '0', md: '6' }}>
+            <Spacer />
+            <HStack>
                 <IconButton
                     size="lg"
                     variant="ghost"
@@ -62,12 +58,13 @@ const MobileNav = ({ onOpen, userInfo, workspace, ...rest }: MobileProps) => {
                 <Button onClick={toggleColorMode} data-cy="switchThemeButton" mr="1%">
                     {colorMode === 'light' && <MoonIcon /> || <SunIcon />}
                 </Button>
-                <Flex alignItems={'center'}>
+                <Flex>
                     <Menu>
                         <MenuButton
                             py={2}
                             transition="all 0.3s"
-                            _focus={{ boxShadow: 'none' }}>
+                            _focus={{ boxShadow: 'none' }}
+                        >
                             <HStack>
                                 <Avatar
                                     size={'sm'}
@@ -76,7 +73,7 @@ const MobileNav = ({ onOpen, userInfo, workspace, ...rest }: MobileProps) => {
                                     }
                                 />
                                 <VStack
-                                    display={{ base: 'none', md: 'flex' }}
+                                    display={'flex'}
                                     alignItems="flex-start"
                                     spacing="1px"
                                     ml="2">
@@ -85,7 +82,7 @@ const MobileNav = ({ onOpen, userInfo, workspace, ...rest }: MobileProps) => {
                                         {role}
                                     </Text>
                                 </VStack>
-                                <Box display={{ base: 'none', md: 'flex' }}>
+                                <Box display={'flex'}>
                                     <FiChevronDown />
                                 </Box>
                             </HStack>
@@ -93,15 +90,14 @@ const MobileNav = ({ onOpen, userInfo, workspace, ...rest }: MobileProps) => {
                         <MenuList
                             borderColor={useColorModeValue('gray.200', 'gray.700')}>
                             <MenuItem>Profile</MenuItem>
-                            <MenuItem>Settings</MenuItem>
-                            <MenuItem>Billing</MenuItem>
+                            <MenuItem onClick={() => { Router.push(ACCOUNT) }}>Settings</MenuItem>
                             <MenuDivider />
-                            <MenuItem>Sign out</MenuItem>
+                            <MenuItem onClick={() => { api.post(SECURE_AUTH_SIGNOUT).then(() => Router.push(ROOT)); }}>Sign out</MenuItem>
                         </MenuList>
                     </Menu>
                 </Flex>
             </HStack>
-        </Flex>
+        </Flex >
     );
 };
-export default MobileNav;
+export default NavBar;